This closes bug #507
If e17 is compiled with tslib but during runtime the environment variable
TSLIB_TSDEVICE is not set, ecore will accidentally use the file descriptor 0
(probably stdin) for tslib.

The problem is that _ecore_fb_ts_fd is initialized to 0, which is BAD BAD BAD
for file descriptors. It should be initialized to -1. The attached patch fixes
this. 

Thanks to John Ogness <john.ogness@linutronix.de> for bug report and fix




git-svn-id: http://svn.enlightenment.org/svn/e/trunk/ecore@45703 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33